+++ src/containers/Header/Header.js
@@ -4,8 +4,17 @@ import { Link } from 'react-router';
 import AppBar from 'material-ui/AppBar';
 import FlatButton from 'material-ui/FlatButton';
 
+import IconButton from 'material-ui/IconButton';
+import IconMenu from 'material-ui/IconMenu';
+import MenuItem from 'material-ui/MenuItem';
+import MoreVertIcon from 'material-ui/svg-icons/navigation/more-vert';
+
 import './Header.scss';
 
+const iconMenuStyle = {
+	position: 'relative',
+	top: '-16px',
+};
 
 class Header extends React.Component {
 	render () {
@@ -15,14 +24,27 @@ class Header extends React.Component {
 		};
 		const homeLink = <Link to="/" activeClassName={'active'} />;
 		const compareLink = <Link to="/compare" activeClassName={'active'} />;
+		const compareByShopsLink = <Link to="/compareByShops" activeClassName={'active'} />;
 		const aboutLink = <Link to="/about" activeClassName={'active'} />;
 		const demoLink = <Link to="/demo" activeClassName={'active'} />;
+
+		const dropdown = (
+			<IconMenu
+				style={iconMenuStyle}
+				iconButtonElement={<IconButton><MoreVertIcon color="#ffffff" /></IconButton>}
+				targetOrigin={{horizontal: 'right', vertical: 'top'}}
+				anchorOrigin={{horizontal: 'right', vertical: 'top'}}>
+				<MenuItem primaryText="Compare two Shops" linkButton containerElement={compareLink} />
+				<MenuItem primaryText="Compare Consumption by Shops" linkButton containerElement={compareByShopsLink} />
+			</IconMenu>
+		);
+
 		const rightButtons = (
 			<nav>
 				<FlatButton style={buttonStyle} containerElement={homeLink} label="Home" />
-				<FlatButton style={buttonStyle} containerElement={compareLink} label="Compare" />
 				<FlatButton style={buttonStyle} containerElement={aboutLink} label="About" />
 				<FlatButton style={buttonStyle} containerElement={demoLink} label="Demo" />
+				{dropdown}
 			</nav>
 		);